fix: consolidate variables page into secrets page
Remove the duplicate /settings/variables page and route module; env vars
are secrets, so managing them in two places was confusing. Port the
variables page's partial-value masking (last 4 chars) into secrets.py/html,
and add a 301 redirect from /settings/variables to /settings/secrets for
existing bookmarks.
