+++ dango/web/app.py
@@ -546,7 +546,6 @@ from dango.web.routes.sync import router as sync_router  # noqa: E402
 from dango.web.routes.ui import router as ui_router  # noqa: E402
 from dango.web.routes.upload import router as upload_router  # noqa: E402
 from dango.web.routes.users import router as users_router  # noqa: E402
-from dango.web.routes.variables import router as variables_router  # noqa: E402
 from dango.web.routes.websocket import router as websocket_router  # noqa: E402
 
 # Dango API routers (order matters — more specific routes first)
@@ -561,7 +560,6 @@ app.include_router(initial_sync_router)
 app.include_router(logs_router)
 app.include_router(upload_router)
 app.include_router(secrets_router)
-app.include_router(variables_router)
 app.include_router(oauth_connect_router)
 app.include_router(dbt_router)
 app.include_router(schedules_router)
