+++ dango/web/templates/secrets.html
@@ -54,7 +54,7 @@
                         <template x-for="v in envVars" :key="v.key">
                             <tr>
                                 <td class="px-6 py-4 whitespace-nowrap text-sm font-mono text-gray-900" x-text="v.key"></td>
-                                <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">***</td>
+                                <td class="px-6 py-4 whitespace-nowrap text-sm font-mono text-gray-600" x-text="v.masked_value"></td>
                                 <td class="px-6 py-4 whitespace-nowrap text-right text-sm space-x-2">
                                     <button @click="openEditModal(v.key)"
                                             class="text-dango-600 hover:text-dango-800 font-medium">Edit</button>
